backend/requirements.txt
fastapi
uvicorn
sqlmodel
passlib
python-jose
scikit-learn
joblib
pytesseract
reportlab
pillow  
requests
email-validator